requirements.txt
pandas==1.2.2
ilock==1.0.3
gitpython==2.2.1
docopt==0.6.2